refactor/fix: resolve circular dependency in extraction modules
Move save_result function to new file_utils module to resolve circular import
between extract_procedure.py and prompt_chain.py. This improves code organization
and maintainability by:
- Create src/lib/file_utils.py for file operation utilities
- Add proper type hints using Union from typing
- Update imports in both affected modules

The change maintains all existing functionality while eliminating the circular
dependency, allowing the extraction pipeline to run without import errors.
